requirements.txt
flask==2.0.1
flask-mail==0.9.1
flask-limiter==2.4.0
flask-caching==2.0.2
python-dotenv==0.19.0
captcha==0.4
redis==4.5.4
gunicorn==20.1.0 